Correalation
Correlations  
 
[1.] Select a professional organization that you are aware of or one that you fancy or 
support. Visit their web site in an attempt to search for facts and figures about the 
organization, its activities, affiliates or stakeholders.  
¾ Hints: 
§ It could be a sports organization, but other types are also possible 
§ The facts and figures could be collected about athletes, in the case 
of sports organization 
[2.] The datasets which you are looking for could be readily available or that you have 
to prepare it. 
¾ Hint: 
§ See the appendix for potential dataset repositories  
[3.] In  case  dataset  is  not  readily  available,  you  will  have  to  prepare  it.  that  could  be 
done in various ways e.g., Excel or OpenOffice Calc.  
¾ Hints:  
§ In  the  case  of  sports  organization,  you  may  wish  to  open  either 
Excel or Calc and starting in Cell A across Row 1 of the spreadsheet, 
define some attributes (at least three or four) to hold data about 
each athlete 
§ Some possible attributes you may wish to consider could be salary, 
goals,  experience,  height,  weight,  age,  etc.  The  list  is  potentially 
unlimited, will vary based on the type of sport you choose, and will 
depend  on  the  data  available  to  you  on  the  web  site(s)  you  have 
selected  
§ Measurements  of  the  athletes’  salaries  and  performance  in 
competition are likely to be the most interesting 
§ Other attributes such as athletes’ names could be added. However, 
keep in mind that correlations can only be conducted on numeric 
data! accordingly, the name attribute needs to be reduced out of 
your  dataset  before  creating  the  correlation  matrix.  Use 
RapidMiner Select Attributes Operator for that purpose! 
[4.] Look  up  the  statistics  for  each  of  your  selected  attributes  and  enter  them  as 
observations into your spreadsheet. Try to find as many as you can 
¾ Hints: 
§ Normally 100 works as a good rule of thumb in order to achieve at 
least a basic level of statistical validity  
§ More is better!

[5.] Once you have created your dataset, save it as a CSV file. Click File, then Save As. 
Enter a file name, and change ‘Save as type:’ to be Text CSV (.csv). Be sure to save 
the file in your data mining data (sub)folder 
¾ Hint: 
§ The file could also be saved as .xlsx 
[6.] Open RapidMiner and import your dataset into your RapidMiner repository. 
Name  it  assignment1,  or  something  descriptive  so  that  you  will  remember  what 
data are contained in the dataset when you look in your repository 
¾ Hints: 
§ Import .csv file using RapidMiner Read CSV Operator 
§ Import .xlsx file using RapidMiner Read Excel Operator 
[7.] Add  the  dataset  to  a  new  process  in  RapidMiner.  Ensure  that  the  out  port  is 
connected to a res port and run your model. SAVE your process with a descriptive 
name if you wish. Examine your data in RESULTS perspective and ensure there are 
no missing, inconsistent, or other potentially problematic data that might need to 
be handled as part of your data preparation phase. Return to DESIGN perspective 
and handle any data preparation tasks that may be necessary 
[8.] Add the Correlation Matrix Operator to your stream and ensure that the mat 
port is connected to a res port. RUN your model again. Interpret your correlation 
coefficients as displayed on the matrix tab 
[9.] Document your findings. In your report, answer the below questions: 
¾ What correlations exist?  
¾ How strong are they?  
¾ Are they surprising to you? If so, why? 
¾ What other attributes would you like to add?  
¾ Are there any you would like to  eliminate now? Why?