Decision trees
[1.] Conduct an Internet search for the Titanic passenger list. The search term ‘Titanic 
passenger  list’  in  your  favorite  search  engine  will  yield  a  number  of  web  sites 
containing lists of passengers 
[2.] Select  from  the  sources  you  find  a  sample  of  passengers.  You  do  not  need  to 
construct a training dataset of every passenger on the Titanic (unless you want to), 
but get at least 300, or preferably more. The more robust your training dataset is, 
the more interesting your results will be 
[3.] In a spreadsheet e.g., Excel or OpenOffice Calc, enter these passengers’ data: 
¾ Hints: 
§ Record attributes such as their name, age, gender, class of service 
they  traveled  in,  background  or  nationality  if  known,  or  other 
attributes  that  may  be  available  to  you  depending  on  the  detail 
level of the data source you find 
§ Be  sure  to  have  at  least  four  attributes,  or  preferably  more. 
Remember that the passengers’ names or ID numbers will not be 
predictive, so that attribute should not be counted as one of your 
predictor attributes 
§ Add to your dataset whether the person lived (was rescued from a 
life  boat  or  from  the  water),  died  (their  body  was  recovered),  or 
was  lost  (was  on  the  Titanic’s  manifest  but  was  never  accounted 
for and therefore presumed dead after the ship’s sinking). Call this 
attribute ‘Survival_Result’ 
§ Save  this  spreadsheet  as  a  CSV  file  and  then  import  it  into  your 
RapidMiner  repository.  Set  the  Survival_Result  attribute’s  role  to 
be your label. Set other attributes which are not predictive, such as 
names, to not be considered in the decision tree model 
§ Add a Decision Tree Operator to your stream 
[4.] If you have searched well, you will be able to find the above dataset read, in case 
you did not, please do all above steps! 
[5.] In  a  new,  blank  spreadsheet,  duplicate  the  attribute  names  from  your  training 
dataset, with the exception of Survival_Result. You will predict this attribute using 
your decision tree  
[6.] Enter data for yourself and people that you know into this spreadsheet 
¾ Hints: 
§ For  some  attributes,  you  may  have  to  decide  what  to  put.  For 
example, the author acknowledges that based on how relentlessly 
he searches for the absolutely cheapest ticket when shopping for 
airfare, he almost certainly would have been in 3rd class if he had 
been on the Titanic. He further knows some people who very likely 
would have been in 1st class 
§ If you want to include some people in your dataset but you do not 
know  every  single  attribute  for  them,  remember,  decision  trees 
can handle some missing values 
§ Save  this  spreadsheet  as  a  .CSV  file  and  import  it  into  your 
RapidMiner repository 
§ Drag this data set into your process and ensure that attributes that 
are not predictive, such as names, will not be included as predictors 
in the model 
[7.] Apply your decision tree model to your scoring dataset 
[8.] Document your findings. In your report, do as per the the below: 
¾ Run  your  model  using  gain_ratio.  Report  your  tree  nodes,  and  discuss 
whether you and the people you know would have lived, died or been lost 
¾ Re-run  your  model  using  gini_index.  Report  differences  in  your  tree 
structures. Discuss whether your chances for survival increase under Gini or 
Gain